requirements.txt
requests
nonebot2
nonebot-adapter-onebot
pillow
pyyaml
lunardate